support-base/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-genai==1.5.0
google-cloud-texttospeech==2.21.1
httpx==0.28.1
pydantic==2.10.4
